Fix HBI value for dihydronaphthalene radical
Species matching this group seemed to have incorrect enthalpies
After comparing with the calculated enthalpy for the species used
to fit the HBI value, it seemed that the value was off by exactly
the enthalpy of an H atom.

Also fix SMILES for the model species used to fit the
Benzyl_S_dihydronaphthalene and Benzyl_T_dihydronaphthalene
group values since they seem to have been swapped.